@@ -1,35 +1,31 @@
 module github.com/Loyalsoldier/geoip
 
-go 1.23
-
-toolchain go1.23.2
+go 1.24
 
 require (
-	github.com/klauspost/compress v1.17.11
+	github.com/klauspost/compress v1.18.0
 	github.com/maxmind/mmdbwriter v1.0.0
-	github.com/oschwald/geoip2-golang v1.11.0
+	github.com/oschwald/geoip2-golang v1.13.0
 	github.com/oschwald/maxminddb-golang v1.13.1
-	github.com/sagernet/sing-box v1.10.1
-	github.com/spf13/cobra v1.8.1
-	github.com/tailscale/hujson v0.0.0-20241010212012-29efb4a0184b
+	github.com/sagernet/sing-box v1.12.4
+	github.com/spf13/cobra v1.10.1
+	github.com/tailscale/hujson v0.0.0-20250605163823-992244df8c5a
 	github.com/tidwall/gjson v1.18.0
 	go4.org/netipx v0.0.0-20231129151722-fdeea329fbba
-	google.golang.org/protobuf v1.35.1
+	google.golang.org/protobuf v1.36.8
 	gopkg.in/yaml.v2 v2.4.0
 )
 
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/kr/text v0.2.0 // indirect
-	github.com/miekg/dns v1.1.62 // indirect
-	github.com/sagernet/sing v0.5.0-rc.2 // indirect
-	github.com/sagernet/sing-dns v0.3.0-rc.2 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/miekg/dns v1.1.67 // indirect
+	github.com/sagernet/sing v0.7.6-0.20250825114712-2aeec120ce28 // indirect
+	github.com/spf13/pflag v1.0.9 // indirect
 	github.com/tidwall/match v1.1.1 // indirect
 	github.com/tidwall/pretty v1.2.0 // indirect
-	golang.org/x/mod v0.19.0 // indirect
-	golang.org/x/net v0.27.0 // indirect
-	golang.org/x/sync v0.8.0 // indirect
-	golang.org/x/sys v0.25.0 // indirect
-	golang.org/x/tools v0.23.0 // indirect
+	golang.org/x/mod v0.27.0 // indirect
+	golang.org/x/net v0.43.0 // indirect
+	golang.org/x/sync v0.16.0 // indirect
+	golang.org/x/sys v0.35.0 // indirect
+	golang.org/x/tools v0.36.0 // indirect
 )
