@@ -13,30 +13,30 @@ import (
 )
 
 const (
-	TypeASNCSV = "maxmindGeoLite2ASNCSV"
-	DescASNCSV = "Convert MaxMind GeoLite2 ASN CSV data to other formats"
+	TypeGeoLite2ASNCSVIn = "maxmindGeoLite2ASNCSV"
+	DescGeoLite2ASNCSVIn = "Convert MaxMind GeoLite2 ASN CSV data to other formats"
 )
 
 var (
-	defaultASNIPv4File = filepath.Join("./", "geolite2", "GeoLite2-ASN-Blocks-IPv4.csv")
-	defaultASNIPv6File = filepath.Join("./", "geolite2", "GeoLite2-ASN-Blocks-IPv6.csv")
+	defaultGeoLite2ASNCSVIPv4File = filepath.Join("./", "geolite2", "GeoLite2-ASN-Blocks-IPv4.csv")
+	defaultGeoLite2ASNCSVIPv6File = filepath.Join("./", "geolite2", "GeoLite2-ASN-Blocks-IPv6.csv")
 )
 
 func init() {
-	lib.RegisterInputConfigCreator(TypeASNCSV, func(action lib.Action, data json.RawMessage) (lib.InputConverter, error) {
-		return newGeoLite2ASNCSV(action, data)
+	lib.RegisterInputConfigCreator(TypeGeoLite2ASNCSVIn, func(action lib.Action, data json.RawMessage) (lib.InputConverter, error) {
+		return newGeoLite2ASNCSVIn(action, data)
 	})
-	lib.RegisterInputConverter(TypeASNCSV, &GeoLite2ASNCSV{
-		Description: DescASNCSV,
+	lib.RegisterInputConverter(TypeGeoLite2ASNCSVIn, &GeoLite2ASNCSVIn{
+		Description: DescGeoLite2ASNCSVIn,
 	})
 }
 
-func newGeoLite2ASNCSV(action lib.Action, data json.RawMessage) (lib.InputConverter, error) {
+func newGeoLite2ASNCSVIn(action lib.Action, data json.RawMessage) (lib.InputConverter, error) {
 	var tmp struct {
-		IPv4File   string              `json:"ipv4"`
-		IPv6File   string              `json:"ipv6"`
-		Want       map[string][]string `json:"wantedList"`
-		OnlyIPType lib.IPType          `json:"onlyIPType"`
+		IPv4File   string                 `json:"ipv4"`
+		IPv6File   string                 `json:"ipv6"`
+		Want       lib.WantedListExtended `json:"wantedList"`
+		OnlyIPType lib.IPType             `json:"onlyIPType"`
 	}
 
 	if len(data) > 0 {
@@ -45,17 +45,17 @@ func newGeoLite2ASNCSV(action lib.Action, data json.RawMessage) (lib.InputConver
 		}
 	}
 
-	if tmp.IPv4File == "" {
-		tmp.IPv4File = defaultASNIPv4File
-	}
-
-	if tmp.IPv6File == "" {
-		tmp.IPv6File = defaultASNIPv6File
+	// When both of IP files are not specified,
+	// it means user wants to use the default ones
+	if tmp.IPv4File == "" && tmp.IPv6File == "" {
+		tmp.IPv4File = defaultGeoLite2ASNCSVIPv4File
+		tmp.IPv6File = defaultGeoLite2ASNCSVIPv6File
 	}
 
 	// Filter want list
-	wantList := make(map[string][]string) // map[asn][]listname
-	for list, asnList := range tmp.Want {
+	wantList := make(map[string][]string) // map[asn][]listname or map[asn][]asn
+
+	for list, asnList := range tmp.Want.TypeMap {
 		list = strings.ToUpper(strings.TrimSpace(list))
 		if list == "" {
 			continue
@@ -76,14 +76,19 @@ func newGeoLite2ASNCSV(action lib.Action, data json.RawMessage) (lib.InputConver
 		}
 	}
 
-	if len(wantList) == 0 {
-		return nil, fmt.Errorf("❌ [type %s | action %s] wantedList must be specified in config", TypeASNCSV, action)
+	for _, asn := range tmp.Want.TypeSlice {
+		asn = strings.TrimPrefix(strings.ToLower(strings.TrimSpace(asn)), "as")
+		if asn == "" {
+			continue
+		}
+
+		wantList[asn] = []string{"AS" + asn}
 	}
 
-	return &GeoLite2ASNCSV{
-		Type:        TypeASNCSV,
+	return &GeoLite2ASNCSVIn{
+		Type:        TypeGeoLite2ASNCSVIn,
 		Action:      action,
-		Description: DescASNCSV,
+		Description: DescGeoLite2ASNCSVIn,
 		IPv4File:    tmp.IPv4File,
 		IPv6File:    tmp.IPv6File,
 		Want:        wantList,
@@ -91,7 +96,7 @@ func newGeoLite2ASNCSV(action lib.Action, data json.RawMessage) (lib.InputConver
 	}, nil
 }
 
-type GeoLite2ASNCSV struct {
+type GeoLite2ASNCSVIn struct {
 	Type        string
 	Action      lib.Action
 	Description string
@@ -101,19 +106,19 @@ type GeoLite2ASNCSV struct {
 	OnlyIPType  lib.IPType
 }
 
-func (g *GeoLite2ASNCSV) GetType() string {
+func (g *GeoLite2ASNCSVIn) GetType() string {
 	return g.Type
 }
 
-func (g *GeoLite2ASNCSV) GetAction() lib.Action {
+func (g *GeoLite2ASNCSVIn) GetAction() lib.Action {
 	return g.Action
 }
 
-func (g *GeoLite2ASNCSV) GetDescription() string {
+func (g *GeoLite2ASNCSVIn) GetDescription() string {
 	return g.Description
 }
 
-func (g *GeoLite2ASNCSV) Input(container lib.Container) (lib.Container, error) {
+func (g *GeoLite2ASNCSVIn) Input(container lib.Container) (lib.Container, error) {
 	entries := make(map[string]*lib.Entry)
 
 	if g.IPv4File != "" {
@@ -158,7 +163,7 @@ func (g *GeoLite2ASNCSV) Input(container lib.Container) (lib.Container, error) {
 	return container, nil
 }
 
-func (g *GeoLite2ASNCSV) process(file string, entries map[string]*lib.Entry) error {
+func (g *GeoLite2ASNCSVIn) process(file string, entries map[string]*lib.Entry) error {
 	if entries == nil {
 		entries = make(map[string]*lib.Entry)
 	}
@@ -193,16 +198,36 @@ func (g *GeoLite2ASNCSV) process(file string, entries map[string]*lib.Entry) err
 			return fmt.Errorf("❌ [type %s | action %s] invalid record: %v", g.Type, g.Action, record)
 		}
 
-		if listArr, found := g.Want[strings.TrimSpace(record[1])]; found {
-			for _, listName := range listArr {
-				entry, got := entries[listName]
-				if !got {
-					entry = lib.NewEntry(listName)
+		// Maxmind ASN CSV reference:
+		// network,autonomous_system_number,autonomous_system_organization
+		// 1.0.0.0/24,13335,CLOUDFLARENET
+		// 1.0.4.0/22,38803,"Gtelecom Pty Ltd"
+		// 1.0.16.0/24,2519,"ARTERIA Networks Corporation"
+
+		switch len(g.Want) {
+		case 0: // it means user wants all ASNs
+			asn := "AS" + strings.TrimSpace(record[1]) // default list name is in "AS12345" format
+			entry, got := entries[asn]
+			if !got {
+				entry = lib.NewEntry(asn)
+			}
+			if err := entry.AddPrefix(strings.TrimSpace(record[0])); err != nil {
+				return err
+			}
+			entries[asn] = entry
+
+		default: // it means user wants specific ASNs or customized lists with specific ASNs
+			if listArr, found := g.Want[strings.TrimSpace(record[1])]; found {
+				for _, listName := range listArr {
+					entry, got := entries[listName]
+					if !got {
+						entry = lib.NewEntry(listName)
+					}
+					if err := entry.AddPrefix(strings.TrimSpace(record[0])); err != nil {
+						return err
+					}
+					entries[listName] = entry
 				}
-				if err := entry.AddPrefix(strings.TrimSpace(record[0])); err != nil {
-					return err
-				}
-				entries[listName] = entry
 			}
 		}
 	}
